# adminPanel/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Education Platform Admin API",
    "description": "API для управления категориями, курсами и уроками",
    "version": "1.0.0"
  },
  "host": "localhost:4000",
  "basePath": "/api/v1",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "OAuth2": {
      "type": "oauth2",
      "flow": "authorizationCode",
      "authorizationUrl": "http://localhost:8080/realms/teacher/protocol/openid-connect/auth",
      "tokenUrl": "http://localhost:8080/realms/teacher/protocol/openid-connect/token",
      "scopes": {
        "openid": "OpenID Connect scope",
        "profile": "User profile information",
        "email": "User email"
      }
    }
  },
  "security": [
    {
      "OAuth2": [
        "openid",
        "profile",
        "email"
      ]
    }
  ],
...
  "paths": {
...
    "/categories/{category_id}/courses": {
      "get": {
        "tags": [
          "Courses"
        ],
        "summary": "Получить курсы категории",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "page",
            "in": "query",
            "type": "integer",
            "default": 1,
            "minimum": 1,
            "description": "Номер страницы"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "integer",
            "default": 20,
            "minimum": 1,
            "maximum": 100,
            "description": "Количество элементов на странице"
          }
        ],
        "responses": {
          "200": {
            "description": "Успешно получены курсы категории",
            "schema": {
              "$ref": "#/definitions/PaginatedCoursesResponse"
            }
          },
          "400": {
            "description": "Неверные параметры запроса",
            "schema": {
              "$ref": "#/definitions/ErrorInvalidUUIDResponse"
            },
            "examples": {
              "application/json": {
                "status": "error",
                "error": {
                  "code": "INVALID_UUID",
                  "message": "Invalid category ID format"
                }
              }
            }
          },
          "404": {
            "description": "Категория не найдена",
            "schema": {
              "$ref": "#/definitions/ErrorNotFoundResponse"
            },
            "examples": {
              "application/json": {
                "status": "error",
                "error": {
                  "code": "NOT_FOUND",
                  "message": "Category not found"
                }
              }
            }
          },
          "500": {
            "description": "Ошибка сервера",
            "schema": {
              "$ref": "#/definitions/ErrorServerErrorResponse"
            },
            "examples": {
              "application/json": {
                "status": "error",
                "error": {
                  "code": "INTERNAL_SERVER_ERROR",
                  "message": "An unexpected error occurred"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Courses"
        ],
        "summary": "Создать новый курс в категории",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CourseCreate"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Курс успешно создан",
            "schema": {
              "$ref": "#/definitions/CourseResponse"
            }
          },
          "400": {
            "description": "Неверные данные",
            "schema": {
              "$ref": "#/definitions/ErrorInvalidUUIDResponse"
            },
            "examples": {
              "application/json": {
                "status": "error",
                "error": {
                  "code": "INVALID_UUID",
                  "message": "Invalid category ID format"
                }
              }
            }
          },
          "404": {
            "description": "Категория не найдена",
            "schema": {
              "$ref": "#/definitions/ErrorNotFoundResponse"
            },
            "examples": {
              "application/json": {
                "status": "error",
                "error": {
                  "code": "NOT_FOUND",
                  "message": "Category not found"
                }
              }
            }
          },
          "500": {
            "description": "Ошибка сервера",
            "schema": {
              "$ref": "#/definitions/ErrorServerErrorResponse"
            },
            "examples": {
              "application/json": {
                "status": "error",
                "error": {
                  "code": "INTERNAL_SERVER_ERROR",
                  "message": "An unexpected error occurred"
                }
              }
            }
          }
        }
      }
    },
    "/categories/{category_id}/courses/{course_id}": {
      "get": {
        "tags": [
          "Courses"
        ],
        "summary": "Получить курс по ID в категории",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Курс найден",
            "schema": {
              "$ref": "#/definitions/CourseResponse"
            }
          },
          "400": {
            "description": "Неверный формат ID",
            "schema": {
              "$ref": "#/definitions/ErrorInvalidUUIDResponse"
            },
            "examples": {
              "application/json": {
                "status": "error",
                "error": {
                  "code": "INVALID_UUID",
                  "message": "Invalid category ID format"
                }
              }
            }
          },
          "404": {
            "description": "Курс или категория не найдена",
            "schema": {
              "$ref": "#/definitions/ErrorNotFoundResponse"
            },
            "examples": {
              "application/json": {
                "status": "error",
                "error": {
                  "code": "NOT_FOUND",
                  "message": "Course not found"
                }
              }
            }
          },
          "500": {
            "description": "Ошибка сервера",
            "schema": {
              "$ref": "#/definitions/ErrorServerErrorResponse"
            },
            "examples": {
              "application/json": {
                "status": "error",
                "error": {
                  "code": "INTERNAL_SERVER_ERROR",
                  "message": "An unexpected error occurred"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Courses"
        ],
        "summary": "Обновить курс в категории",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CourseUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Курс успешно обновлен",
            "schema": {
              "$ref": "#/definitions/CourseResponse"
            }
          },
          "400": {
            "description": "Неверные данные",
            "schema": {
              "$ref": "#/definitions/ErrorInvalidUUIDResponse"
            },
            "examples": {
              "application/json": {
                "status": "error",
                "error": {
                  "code": "INVALID_UUID",
                  "message": "Invalid category ID format"
                }
              }
            }
          },
          "404": {
            "description": "Курс или категория не найдена",
            "schema": {
              "$ref": "#/definitions/ErrorNotFoundResponse"
            },
            "examples": {
              "application/json": {
                "status": "error",
                "error": {
                  "code": "NOT_FOUND",
                  "message": "Course not found"
                }
              }
            }
          },
          "500": {
            "description": "Ошибка сервера",
            "schema": {
              "$ref": "#/definitions/ErrorServerErrorResponse"
            },
            "examples": {
              "application/json": {
                "status": "error",
                "error": {
                  "code": "INTERNAL_SERVER_ERROR",
                  "message": "An unexpected error occurred"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Courses"
        ],
        "summary": "Удалить курс из категории",
        "description": "Удаляет курс и все связанные уроки",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Курс успешно удален",
            "schema": {
              "$ref": "#/definitions/StatusOnly"
            }
          },
          "400": {
            "description": "Неверный формат ID",
            "schema": {
              "$ref": "#/definitions/ErrorInvalidUUIDResponse"
            },
            "examples": {
              "application/json": {
                "status": "error",
                "error": {
                  "code": "INVALID_UUID",
                  "message": "Invalid category ID format"
                }
              }
            }
          },
          "404": {
            "description": "Курс или категория не найдена",
            "schema": {
              "$ref": "#/definitions/ErrorNotFoundResponse"
            },
            "examples": {
              "application/json": {
                "status": "error",
                "error": {
                  "code": "NOT_FOUND",
                  "message": "Course not found"
                }
              }
            }
          },
          "500": {
            "description": "Ошибка сервера",
            "schema": {
              "$ref": "#/definitions/ErrorServerErrorResponse"
            },
            "examples": {
              "application/json": {
                "status": "error",
                "error": {
                  "code": "INTERNAL_SERVER_ERROR",
                  "message": "An unexpected error occurred"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "StatusOnly": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success",
          "description": "Результат выполнения"
        }
      }
    },
    "Pagination": {
      "type": "object",
      "properties": {
        "total": {
          "type": "integer",
          "example": 150,
          "description": "Общее количество элементов"
        },
        "page": {
          "type": "integer",
          "example": 1,
          "description": "Текущая страница"
        },
        "limit": {
          "type": "integer",
          "example": 20,
          "description": "Количество элементов на странице"
        },
        "pages": {
          "type": "integer",
          "example": 8,
          "description": "Общее количество страниц"
        }
      }
    },
...
    "Course": {
      "type": "object",
      "description": "Курс",
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid",
          "example": "660e8400-e29b-41d4-a716-446655440001",
          "description": "Уникальный идентификатор"
        },
        "title": {
          "type": "string",
          "maxLength": 255,
          "example": "Go для начинающих",
          "description": "Название курса"
        },
        "description": {
          "type": "string",
          "example": "Изучите Golang с нуля",
          "description": "Описание курса"
        },
        "level": {
          "type": "string",
          "enum": [
            "hard",
            "medium",
            "easy"
          ],
          "example": "medium",
          "description": "Уровень сложности",
          "default": "medium"
        },
        "category_id": {
          "type": "string",
          "format": "uuid",
          "example": "550e8400-e29b-41d4-a716-446655440000",
          "description": "ID категории"
        },
        "visibility": {
          "type": "string",
          "enum": [
            "draft",
            "public",
            "private"
          ],
          "example": "draft",
          "description": "Видимость курса",
          "default": "draft"
        },
        "created_at": {
          "type": "string",
          "format": "date-time",
          "example": "2024-01-15T11:00:00Z",
          "description": "Дата создания"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time",
          "example": "2024-01-15T11:00:00Z",
          "description": "Дата обновления"
        }
      }
    },
    "CourseCreate": {
      "type": "object",
      "required": [
        "title",
        "category_id"
      ],
      "properties": {
        "title": {
          "type": "string",
          "maxLength": 255,
          "example": "Go для начинающих",
          "description": "Название курса"
        },
        "description": {
          "type": "string",
          "example": "Изучите Golang с нуля",
          "description": "Описание курса"
        },
        "level": {
          "type": "string",
          "enum": [
            "hard",
            "medium",
            "easy"
          ],
          "example": "medium",
          "description": "Уровень сложности",
          "default": "medium"
        },
        "category_id": {
          "type": "string",
          "format": "uuid",
          "example": "550e8400-e29b-41d4-a716-446655440000",
          "description": "ID категории"
        },
        "visibility": {
          "type": "string",
          "enum": [
            "draft",
            "public",
            "private"
          ],
          "example": "draft",
          "description": "Видимость курса",
          "default": "draft"
        }
      }
    },
    "CourseUpdate": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "maxLength": 255,
          "example": "Golang для профессионалов",
          "description": "Название курса"
        },
        "description": {
          "type": "string",
          "example": "Продвинутый курс по Go",
          "description": "Описание курса"
        },
        "level": {
          "type": "string",
          "enum": [
            "hard",
            "medium",
            "easy"
          ],
          "example": "hard",
          "description": "Уровень сложности"
        },
        "category_id": {
          "type": "string",
          "format": "uuid",
          "example": "550e8400-e29b-41d4-a716-446655440000",
          "description": "ID категории"
        },
        "visibility": {
          "type": "string",
          "enum": [
            "draft",
            "public",
            "private"
          ],
          "example": "public",
          "description": "Видимость курса"
        }
      }
    },
...
    "PaginatedCoursesResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "data": {
          "type": "object",
          "properties": {
            "items": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Course"
              },
              "description": "Список курсов"
            },
            "pagination": {
              "$ref": "#/definitions/Pagination"
            }
          }
        }
      }
    },
    "CourseResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "data": {
          "$ref": "#/definitions/Course",
          "description": "Данные курса"
        }
      }
    },
...
    "ErrorInvalidUUIDResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "error"
        },
        "error": {
          "type": "object",
          "properties": {
            "code": {
              "type": "string",
              "example": "INVALID_UUID",
              "description": "Неверный формат UUID"
            },
            "message": {
              "type": "string",
              "example": "Invalid ID format",
              "description": "Описание ошибки"
            }
          }
        }
      }
    },
    "ErrorNotFoundResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "error"
        },
        "error": {
          "type": "object",
          "properties": {
            "code": {
              "type": "string",
              "example": "NOT_FOUND",
              "description": "Ресурс не найден"
            },
            "message": {
              "type": "string",
              "example": "Resource not found",
              "description": "Описание ошибки"
            }
          }
        }
      }
    },
    "ErrorServerErrorResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "error"
        },
        "error": {
          "type": "object",
          "properties": {
            "code": {
              "type": "string",
              "example": "SERVER_ERROR",
              "description": "Внутренняя ошибка сервера"
            },
            "message": {
              "type": "string",
              "example": "Internal server error",
              "description": "Описание ошибки"
            }
          }
        }
      }
    },
...
  }
}