# publicSide/internal/dto/response/course.go
package response

import "time"

// CourseDTO represents a course data transfer object.
type CourseDTO struct {
	ID          string    `json:"id"`
	Title       string    `json:"title"`
	Description string    `json:"description"`
	Level       string    `json:"level"`
	CategoryID  string    `json:"category_id"`
	ImageURL    string    `json:"image_url"` // Added ImageURL
	CreatedAt   time.Time `json:"created_at"`
	UpdatedAt   time.Time `json:"updated_at"`
}